Update Tap Detector output type

+++ DsfLogger.cpp
@@ -597,7 +597,7 @@ void DsfLogger::logSensorValue(sh2_SensorValue_t* pValue, double currTime) {
             }
             LogReportCommon(id, currTime);
             outFile_ << extender.extend(pValue->sequence) << ",";
-            outFile_ << pValue->un.tapDetector.flags << ",";
+            outFile_ << static_cast<uint32_t>(pValue->un.tapDetector.flags) << ",";
             outFile_ << static_cast<uint32_t>(pValue->status) << "\n";
             break;
         }
